backend/requirements.txt
pandas==1.5.2
bresenham==0.2.1
more-itertools==8.10.0
simplejson==3.17.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
Flask-Migrate==0.0.0
urllib3==1.26.5
SQLAlchemy==1.4.45
python-dotenv==0.21.0
jwt==1.3.1
psycopg2-binary==0.0.0
